updated book.md & depend files
This contains today's update of book.md and dependency files (word count and histogram plots), both as html and .MD. I've updated the general language to be accurate with respect to type I/II and the data prep., based on @pbailey update of the suppressed data earlier today. Without his counts, I've removed the section of counts from the suppression section of this.